Added styled-components and better styleguide config
