+++ package.json
@@ -27,10 +27,12 @@
     "@types/node": "^11.13.0",
     "@types/react": "^16.8.12",
     "@types/react-dom": "^16.8.3",
+    "@types/styled-components": "^4.1.13",
     "babel-loader": "^8.0.5",
     "codecov": "^3.3.0",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.7.1",
+    "react-docgen-typescript": "^1.12.3",
     "react-styleguidist": "^9.0.5",
     "react-testing-library": "^6.1.1",
     "ts-jest": "^24.0.1",
@@ -41,7 +43,8 @@
   },
   "dependencies": {
     "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "react-dom": "^16.8.6",
+    "styled-components": "^4.2.0"
   },
   "jest": {
     "transform": {
