+++ src/components/Loader/styled/AnimatedSvg.ts
@@ -1,0 +1,16 @@
+import styled, { keyframes } from 'styled-components'
+
+const infiniteSpin = keyframes`
+  from {transform: rotate(0deg)}
+  to {transform: rotate(360deg)}
+`
+
+const AnimatedSvg = styled.svg`
+  transition-property: transform;
+  animation-name: ${infiniteSpin};
+  animation-iteration-count: infinite;
+  animation-timing-function: linear;
+`
+
+/** @component */
+export default AnimatedSvg
