+++ src/components/Loader/typings.d.ts
@@ -1,0 +1,10 @@
+type Measure = 'sm' | 'md' | 'lg' | 'xl'
+type Speed = 'fast' | 'slow' | 'normal'
+
+interface LoaderProps {
+  color: string
+  speed: Speed
+  gap: Measure
+  thickness: Measure
+  size: string
+}
