+++ styleguide.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  title: 'React Style Guide Example',
+  components: 'src/components/**/index.tsx',
+  propsParser: require('react-docgen-typescript').withCustomConfig(
+    './tsconfig.json'
+  ).parse,
+}
