+++ tsconfig.json
@@ -11,5 +11,5 @@
     "isolatedModules": true,
     "esModuleInterop": true
   },
-  "include": ["src"]
+  "include": ["./**/*.d.ts", "./**/*.tsx"]
 }
